requirements.txt
# ============================================================
# Financial Analyzer — Python Dependencies
# Install with: pip install -r requirements.txt
# ============================================================

# Web framework
flask>=3.0.0
werkzeug>=3.0.0
gunicorn>=21.0.0          # Production WSGI server (use instead of flask dev server)

# Data & finance
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.36
scipy>=1.11.0
openpyxl>=3.1.0           # Excel file reading/writing

# PDF processing
pdfplumber>=0.10.0
Pillow>=10.0.0

# OCR (for scanned PDFs)
pytesseract>=0.3.10
opencv-python>=4.8.0      # Image preprocessing for OCR

# HTTP requests (price proxy fallbacks)
requests>=2.31.0
